pyproject.toml
[tool.poetry]
name = "tap-s3"
version = "0.1.0"
description = ""
authors = ["Michael Cooper <macoop2363@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.16.59"
python-dotenv = "^0.15.0"
singer-python = "^5.9.1"
pandas = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
